requirements.txt
anyio==3.6.2
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
ecdsa==0.18.0
exceptiongroup==1.0.0rc9
fastapi==0.85.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyhumps==3.8.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
